requirements.txt
Pillow==8.2.0
praw==7.2.0
prawcore==2.0.0
requests==2.25.1
